Move Cobs Wire Worker behind feature flag.
